refactor: extract admin notice hook
