+++ frontend/src/pages/dashboard/admin/payments/index.js
@@ -31,24 +31,7 @@ import { sendChatMessage } from '@/services/messageService';
 import useAuthStore from '@/store/auth/authStore';
 import useNotificationStore from '@/store/notifications/notificationStore';
 import useMessageStore from '@/store/messages/messageStore';
-
-const useAdminNotice = () => {
-  const user = useAuthStore((s) => s.user);
-  const refreshNotifications = useNotificationStore((s) => s.fetch);
-  const refreshMessages = useMessageStore((s) => s.fetch);
-  return async (type, message) => {
-    try {
-      await createNotification({ user_id: user.id, type, message });
-      await sendChatMessage(user.id, { text: message });
-      refreshNotifications?.();
-      refreshMessages?.();
-    } catch (err) {
-      console.error(err);
-      const msg = err.response?.data?.message || 'Failed to send notification';
-      toast.error(msg);
-    }
-  };
-};
+import useAdminNotice from '@/hooks/useAdminNotice';
 
 
 const defaultConfig = {
